+++ Changelog.md
@@ -1,4 +1,10 @@
+12/06/2022
 
+* New functions  TNovusRODBSessionManager class function CreateSession(const aSessionID: TGUID): TROSession; and procedure CreateSession2(aSession: TROSession);
+
+2/12/2021
+
+* New function TNovusROOlympiaSessionManager.CheckSessionIsExists(aSessionID: String) : boolean;
 * Updated for Remobjects 10.x Remoting SDK and Data Abstract 
 * Updated for Delphi 10.3 and Delphi 10.4
 * Removed support for Delphi X4 - X8